Update pyproject.toml to help stupid poetry restrction
Yes, I think this is stupid. You can't future proof against everything, and it's not that it's not worth trying, it's just not worth PREDICTING when there isn't even such a capability forecast. If I don't specify the cap I'm willing to upgrade to, and I upgrade, then that is my fault. Don't crash on install because you think I might one day possibly do something silly based on what another package developer tried to do.